requirements.txt
google-cloud-bigquery==2.22.1
google-cloud-bigquery-storage==2.1.0
google-cloud-error-reporting==1.0.0
google-cloud-logging==1.15.1
google-cloud-storage==1.37.1
google-cloud-secret-manager==2.0.0
pandas_gbq==0.14.1
pyarrow==2.0.0
pandas==1.1.4
Salesforce-FuelSDK==1.3.0
pysftp==0.2.9
numpy
Flask
gunicorn





